upated week2 assignment to specification
