+++ java/src/BST.java
@@ -13,10 +13,7 @@
  *
  *  Other collaborators: None.
  *  References:
- *  1) https://leetcode.com/problems/convert-sorted-list-to-binary-search-tree/solutions/5791984/on-dsw-soln/ 
- *  2) https://leetcode.com/problems/binary-tree-preorder-traversal/solutions/5690881/o-n-morris-traversal-soln/
- *  3) https://leetcode.com/problems/binary-tree-inorder-traversal/solutions/5629922/o-n-morris-traversal-soln/
- *  4) https://leetcode.com/problems/binary-tree-postorder-traversal/solutions/5704500/o-n-morris-traveral-soln/
+ *  1) https://www.geeksforgeeks.org/binary-tree-data-structure/ 
  *
  *  Assignment No.: 2
  *
@@ -33,38 +30,10 @@
  *
  *  -----------------------------------------------------------------------------------
  *
- *  Table with running times(in nanoseconds) measured for different values of 'n'
- *
- *  First Value          |  n = 10^3   |    n = 10^4   |    n = 10^5    |    n = 10^6      |  n = 10^7
- *
- *  Right skewed BST     |  1000       |    125        |    208         |    416           |  750                  
- *  Balanced BST         |  791        |    416        |    583         |    1500          |  4167                  
- *
- *
  *
  *  Table with running times(in nanoseconds) measured for different values of 'n'
  *
- *  Middle value         |  n = 10^3   |    n = 10^4   |    n = 10^5    |    n = 10^6      |  n = 10^7
- *  
- *  Right skewed BST     |  5292       |    49375      |    413125      |    734625        |  39010209                   
- *  Balanced BST         |  375        |    333        |    500         |    5125          |  3541                  
- *
- *
- *
- *
- *  Table with running times(in nanoseconds) measured for different values of 'n'
- *
- *  Last Value           |  n = 10^3   |    n = 10^4   |    n = 10^5    |    n = 10^6      |  n = 10^7
- *
- *  Right skewed BST     |  10125      |    97417      |    227250      |    1880459       |  21926750                  
- *  Balanced BST         |  542        |    416        |    625         |    1875          |  28500                  
- *
- *
- *
- *
- *  Table with running times(in nanoseconds) measured for different values of 'n'
- *
- *  Value not in Tree    |  n = 10^3   |    n = 10^4   |    n = 10^5    |    n = 10^6      |  n = 10^7
+ *                       |  n = 10^3   |    n = 10^4   |    n = 10^5    |    n = 10^6      |  n = 10^7
  *
  *  Right skewed BST     |  10542      |    97125      |    230042      |    1448583       |  19606750                  
  *  Balanced BST         |  333        |    250        |    333         |    708           |  3333                  
@@ -122,8 +91,15 @@
 import java.util.List;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Random;
 
 public final class BST {
+
+    // class only contains static function no need to create an object of this class
+    // hence private constructor
+    private BST() {
+
+    }
 
     /**
      * A class representing a node in Binary Tree
@@ -153,19 +129,58 @@ public final class BST {
             System.out.println("\nTesting search time of skewed and balancedBST of size -> " + nodeCount + "\n");
             System.out.println("------------------------------------------------------------------------------");
 
-            final TreeNode balancedBST = generateBalancedBST(generateRightSkewedBST(nodeCount), nodeCount);
+            final TreeNode balancedBST = generateBalancedBST(nodeCount);
             final TreeNode rightSkewedBST = generateRightSkewedBST(nodeCount);
 
-            searchFirstValue(rightSkewedBST, balancedBST, 1);
-            searchMiddleValue(rightSkewedBST, balancedBST, nodeCount / 2);
-            searchLastValue(rightSkewedBST, balancedBST, nodeCount);
-            searchValueNotInBST(rightSkewedBST, balancedBST, nodeCount + 1);
+            searchValueInBST(rightSkewedBST, balancedBST, nodeCount + 1);
 
             System.out.println("------------------------------------------------------------------------------");
             System.out.println("\nTesting completed for skewed and balancedBST of size -> " + nodeCount + "\n");
             System.out.println("------------------------------------------------------------------------------");
 
         }
+    }
+
+    private static TreeNode generateBalancedBST(final int nodeCount) {
+        TreeNode root = null;
+        for (final int val : generateUniqueRandomNumbers(nodeCount)) {
+            root = insertIntoBST(root, val);
+        }
+        return root;
+    }
+
+    private static TreeNode insertIntoBST(final TreeNode root, final int val) {
+        if (root == null)
+            return new TreeNode(val);
+
+        var iterator = root;
+
+        while (iterator != null) {
+            if (val < iterator.val) {
+                if (iterator.left == null) {
+                    iterator.left = new TreeNode(val);
+                    return root;
+                }
+                iterator = iterator.left;
+            } else {
+                if (iterator.right == null) {
+                    iterator.right = new TreeNode(val);
+                    return root;
+                }
+                iterator = iterator.right;
+            }
+        }
+        return root;
+    }
+
+    public static List<Integer> generateUniqueRandomNumbers(final int n) {
+        final Random random = new Random();
+        List<Integer> numbers = new ArrayList<>();
+
+        for (int i = 1; i <= n; i++)
+            numbers.add(i * random.nextInt(100000));
+        Collections.shuffle(numbers);
+        return numbers;
     }
 
     /**
@@ -186,148 +201,6 @@ public final class BST {
         }
         return false;
     }
-
-    /**************************************************************************************************
-     ************************** DSW IMPLEMENTATION BEGIN HERE
-     * *****************************************
-     ***************************************************************************************************/
-
-    /****
-     * This function is driver function for the DSW implementation
-     * This taken in a right skewed bst ie Vine or backbone tree then performs
-     * neccessary left rotation make the a balanced bst
-     * This uses a dummy node right pointer as the attachment from which to performs
-     * a left rotation
-     * Since in a balanced BST the different in the height of the tree
-     * cannot exceed 1
-     * So first we get the count of the perfect tree ie we get the node count
-     * of the maximum level that we can fill for instance if the node count is 7
-     * that we can form a perfect tree of level 3 ie all leafs are at the same
-     * height, but that might not be the case since say for the 4 level to be
-     * completed filled we will need 8 more leafs ie log2 (nodecount + 1) = 4
-     * but if we have 10 leaves the last level will have 3 leaves nodes, will the
-     * height diff of the leaves of the bst will 1
-     * This building the level of the tree is done by the function
-     * rotateleftbyamount
-     * this function is run a loop and it take the amount count then is does the
-     * left rotationn by that amount so, each cal to this function build level
-     * by performing the left rotation from the parent and the parents right
-     * and moving the parent to is right node and does the left rotation where
-     * fist we do the left rotation by the difference of total node - perfect node
-     * if this difference is not zero, then the generated bst will have a height
-     * diff of
-     * 1 and then in the while loop will build the perferct tree levels
-     * and since the each time we do the rotation from the dummy and dummy,right
-     * since the balanced bst is attached to dummies right
-     *
-     * @param root
-     * @param nodeCount
-     * @return A balanced BST
-     */
-    private static TreeNode generateBalancedBST(final TreeNode root, final int nodeCount) {
-        final TreeNode dummy = new TreeNode(-1, null, root);
-        var perfectTreeNodeCount = getPerfectTreeNodeCount(nodeCount);
-        rotateLeftByAmount(dummy, nodeCount - perfectTreeNodeCount);
-        while (perfectTreeNodeCount > 1) {
-            perfectTreeNodeCount /= 2;
-            rotateLeftByAmount(dummy, perfectTreeNodeCount);
-        }
-        return dummy.right;
-    }
-
-    /****
-     * calculate the amount of nodes present in the perfect trees for eg
-     * say we have the node count as 10 then this function will first calculate
-     * log2 of the nodecount + 1 since numbers of node in the tree is off by 1
-     * due to fact the root is 1
-     * so first we get the floor of log2 of nodecount + 1 ie log2 floor (11)
-     * and this will be 3 since 2 ^ 3 = 8
-     * so now we now the 3 level of the tree will completed completely
-     * ie level 0, 1 and 3 will have full nodes since the total number of
-     * node in this example is 10
-     * Now we know the max filled level, so now calculate the number of nodes in
-     * these level which will 2 ^ 3 -1 (since root is 1)
-     * so we get the perfect node count as 7
-     * hence the fourth level will have 3 node
-     * so we return the perfect node count ie 7 since we already have the
-     * total number of nodes
-     *
-     * @param count number of nodes in the perfect tree
-     * @return count of the nodes that will be present in the perfect tree
-     */
-    private static int getPerfectTreeNodeCount(final double count) {
-        return (int) Math.pow(2, (Math.floor(Math.log(count + 1) / Math.log(2)))) - 1;
-    }
-
-    /****
-     * This function does the left rotation
-     * the visual for this is
-     * say we have a tree like
-     * ****** dummy
-     * *********** \
-     * ************ node1
-     * ***************** \
-     * ****************** node2
-     * so what is function it holds the dummy node
-     * and push the node1.right ie node2 upwards
-     * the best to think if we take a rorasy bead that is right skewed
-     * and we hold it from top (dummy) and push it from the bottom (node2 )
-     * then the node 1 will be pushed to the left node2
-     *
-     * so applying this function to the above tree now the tree will look
-     * like
-     * ****************dummy
-     * *********************\
-     * **********************node2
-     * ******************** /*****\
-     * ******************node1
-     *
-     * so we attahch the dummyies right to the node2
-     * and node2 left will have node1 and
-     * since and subtree present in the node2 left
-     * will now be attached to node1.right
-     *
-     *
-     * @param parent
-     * @param node
-     */
-    private static void rotateLeft(final TreeNode parent, final TreeNode node) {
-        final TreeNode temp = node.right;
-        node.right = temp.left;
-        temp.left = node;
-        parent.right = temp;
-    }
-
-    /****
-     * this function builds a tree level the amount of rotation done by
-     * this function will generated the same number of nodes that level
-     * for example we have number of nodes as 10 so there will be 3 nodes
-     * in the last level so we will do three left rotation alternatively
-     * ie first will be done from dummy.right
-     * and now dummy will have a new right
-     * so we move down to the right and then do rotation there
-     * and so on..
-     *
-     * @param root   holder root from there the rotation will happen ie
-     *               it right->right will now be its right and its prevous
-     *               right now will be attach to the right->right.left
-     * @param amount the number of left rotations to be performed
-     * 
-     */
-    private static void rotateLeftByAmount(TreeNode root, final int amount) {
-        for (int i = 0; i < amount; i++) {
-            rotateLeft(root, root.right);
-            root = root.right;
-        }
-    }
-
-    /**************************************************************************************************
-     ************************** DSW IMPLEMENTATION END HERE *******************************************
-     ***************************************************************************************************/
-
-    /**************************************************************************************************
-     ************************** UTILITY FUNCTIONS BEGING HERE *****************************************
-     ***************************************************************************************************/
 
     /****
      * This functions generates a right skewed tree with
@@ -367,12 +240,12 @@ public final class BST {
      *
      * @param rightSkewedBST A right skewed BST
      * @param balancedBST    A balancesBST
-     * @param key            Last value of the input sequence
+     * @param key            value to be searched
      */
-    private static void searchLastValue(final TreeNode rightSkewedBST, final TreeNode balancedBST,
+    private static void searchValueInBST(final TreeNode rightSkewedBST, final TreeNode balancedBST,
             final int key) {
         System.out.println("****************************************************\n");
-        System.out.println("Searching last value in Right Skewed BST -> " + key);
+        System.out.println("Searching value in the Right Skewed BST -> " + key);
         var startTime = System.nanoTime();
         var res = iterativeSearch(rightSkewedBST, key);
         var endTime = System.nanoTime();
@@ -380,7 +253,7 @@ public final class BST {
         System.out.println("Value found -> " + res);
         System.out.println("\n****************************************************\n");
 
-        System.out.println("Searching last value in balanced BST-> " + key);
+        System.out.println("Searching value in the Balanced BST-> " + key);
         startTime = System.nanoTime();
         res = iterativeSearch(balancedBST, key);
         endTime = System.nanoTime();
@@ -390,209 +263,13 @@ public final class BST {
 
     }
 
-    /**
-     * This function searched key in the both the BST
-     *
-     * @param rightSkewedBST A right skewed BST
-     * @param balancedBST    A balancesBST
-     * @param key            value not in input the sequence ie (n + 1)
-     */
-    private static void searchValueNotInBST(final TreeNode rightSkewedBST, final TreeNode balancedBST,
-            final int key) {
-        System.out.println("****************************************************\n");
-        System.out.println("Searching a value not present in the Right Skewed BST -> " + key);
-        var startTime = System.nanoTime();
-        var res = iterativeSearch(rightSkewedBST, key);
-        var endTime = System.nanoTime();
-        System.out.println("Time taken in nano seconds -> " + (endTime - startTime));
-        System.out.println("Value found -> " + res);
-        System.out.println("\n****************************************************\n");
-
-        System.out.println("Searching a value not present in the Balanced BST-> " + key);
-        startTime = System.nanoTime();
-        res = iterativeSearch(balancedBST, key);
-        endTime = System.nanoTime();
-        System.out.println("Time taken in nano seconds -> " + (endTime - startTime));
-        System.out.println("Value found -> " + res);
-        System.out.println("\n****************************************************\n");
-
-    }
-
-    /**
-     * This function searched key in the both the BST
-     *
-     * @param rightSkewedBST A right skewed BST
-     * @param balancedBST    A balancesBST
-     * @param key            middle value of the input Sequence
-     */
-    private static void searchMiddleValue(final TreeNode rightSkewedBST, final TreeNode balancedBST,
-            final int key) {
-        System.out.println("****************************************************\n");
-        System.out.println("Searching middle value of input range in Right Skewed BST -> " + key);
-        var startTime = System.nanoTime();
-        var res = iterativeSearch(rightSkewedBST, key);
-        var endTime = System.nanoTime();
-        System.out.println("Time taken in nano seconds -> " + (endTime - startTime));
-        System.out.println("Value found -> " + res);
-        System.out.println("\n****************************************************\n");
-
-        System.out.println("Searching middle value of input range in the Balanced BST-> " + key);
-        startTime = System.nanoTime();
-        res = iterativeSearch(balancedBST, key);
-        endTime = System.nanoTime();
-        System.out.println("Time taken in nano seconds -> " + (endTime - startTime));
-        System.out.println("Value found -> " + res);
-        System.out.println("\n****************************************************\n");
-
-    }
-
-    /**
-     * This function searched key in the both the BST
-     *
-     * @param rightSkewedBST A right skewed BST
-     * @param balancedBST    A balancesBST
-     * @param key            first value of the input Sequence
-     */
-    private static void searchFirstValue(final TreeNode rightSkewedBST, final TreeNode balancedBST,
-            final int key) {
-        System.out.println("****************************************************\n");
-        System.out.println("Searching first value of input range in Right Skewed BST -> " + key);
-        var startTime = System.nanoTime();
-        var res = iterativeSearch(rightSkewedBST, key);
-        var endTime = System.nanoTime();
-        System.out.println("Time taken in nano seconds -> " + (endTime - startTime));
-        System.out.println("Value found -> " + res);
-        System.out.println("\n****************************************************\n");
-
-        System.out.println("Searching first value of input range in the Balanced BST-> " + key);
-        startTime = System.nanoTime();
-        res = iterativeSearch(balancedBST, key);
-        endTime = System.nanoTime();
-        System.out.println("Time taken in nano seconds -> " + (endTime - startTime));
-        System.out.println("Value found -> " + res);
-        System.out.println("\n****************************************************\n");
-
-    }
-
-    /**************************************************************************************************
-     ************************** TREE TRAVERSAL FUNCTIONS BEGIN FROM HERE ******************************
-     ******************** USING MORRIS TRAVERSAL TO OPTIMIZE SPACE COMPLEXCITY ************************
-     * 
-     * USE TRAVERSAL TO CONFIRM DSW IMPLEMENTATION GENERATES BALANCE BST FROM RIGHT
-     * SKEWED TREE
-     *
-     **/
-
-    /****
-     * This function gives the inorder traversal using right threaded binary tree
-     * ie before visting the left node to the left rightmost node and create a link
-     * to return back to current node
-     * morris traversal technique
-     *
-     * @param root root of the bst
-     * @return list containing in-order traversal values
-     */
-    private static List<Integer> InorderTraversal(TreeNode root) {
-        final List<Integer> res = new ArrayList<>();
-        while (root != null) {
-            if (root.left != null) {
-                var pre = root.left;
-                while (pre.right != null && pre.right != root)
-                    pre = pre.right;
-                if (pre.right != null) {
-                    res.add(root.val);
-                    pre.right = null;
-                    root = root.right;
-                } else {
-                    pre.right = root;
-                    root = root.left;
-                }
-            } else {
-                res.add(root.val);
-                root = root.right;
-            }
-        }
-        return res;
-    }
-
-    /****
-     * This function gives the inorder traversal using right threaded binary tree
-     * ie before visting the left node to the left rightmost node and create a link
-     * to return back to current node
-     * morris traversal technique
-     *
-     * @param root root of the bst
-     * @return list containing pre-order traversal values
-     */
-    private static List<Integer> PreorderTraversal(TreeNode root) {
-        final List<Integer> res = new ArrayList<>();
-        while (root != null) {
-            if (root.left != null) {
-                var pre = root.left;
-                while (pre.right != null && pre.right != root)
-                    pre = pre.right;
-                if (pre.right != null) {
-                    pre.right = null;
-                    root = root.right;
-                } else {
-                    res.add(root.val);
-                    pre.right = root;
-                    root = root.left;
-                }
-            } else {
-                res.add(root.val);
-                root = root.right;
-            }
-        }
-        return res;
-    }
-
-    /****
-     * This function gives the postorder traversal using left threaded binary tree
-     * since post order is opposite + reverse of preorder ie
-     * pre order -> root left right
-     * post order -> left right root or reverse of (root right left)
-     * since create left threads ie before visiting the right side
-     * create a thread from its leftmost side to the current node
-     * and reverse the values in the end
-     * morris traversal technique
-     *
-     * @param root root of the bst
-     * @return list containing post-order traversal values
-     */
-    private static List<Integer> PostorderTraversal(TreeNode root) {
-        final List<Integer> res = new ArrayList<>();
-        while (root != null) {
-            if (root.right != null) {
-                var pre = root.right;
-                while (pre.left != null && pre.left != root)
-                    pre = pre.left;
-                if (pre.left != null) {
-                    pre.left = null;
-                    root = root.left;
-                } else {
-                    res.add(root.val);
-                    pre.left = root;
-                    root = root.right;
-                }
-            } else {
-                res.add(root.val);
-                root = root.left;
-            }
-        }
-        Collections.reverse(res);
-        return res;
-    }
-    /**************************************************************************************************
-     ************************** TREE TRAVERSAL FUNCTIONS END HERE *************************************
-     ***************************************************************************************************/
 }
-
-/*******************************************************************************
- * ********************** INPUT AND OUTPUT ************************************
+/*********************************************************
+ ************* INPUT AND OUTPUT***************************
  *
  *
- * ozzy@osaids-MacBook-Air src % java BST 1000 10000 100000 1000000 10000000
+ *
+ * ozzy@Mac src % java BST 1000 10000 100000 1000000 10000000
  * ------------------------------------------------------------------------------
  * 
  * Testing search time of skewed and balancedBST of size -> 1000
@@ -600,56 +277,14 @@ public final class BST {
  * ------------------------------------------------------------------------------
  ****************************************************
  * 
- * Searching first value of input range in Right Skewed BST -> 1
- * Time taken in nano seconds -> 1000
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching first value of input range in the Balanced BST-> 1
- * Time taken in nano seconds -> 791
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching middle value of input range in Right Skewed BST -> 500
- * Time taken in nano seconds -> 5292
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching middle value of input range in the Balanced BST-> 500
- * Time taken in nano seconds -> 375
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching last value in Right Skewed BST -> 1000
- * Time taken in nano seconds -> 10125
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching last value in balanced BST-> 1000
- * Time taken in nano seconds -> 542
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching a value not present in the Right Skewed BST -> 1001
- * Time taken in nano seconds -> 10542
+ * Searching value in the Right Skewed BST -> 1001
+ * Time taken in nano seconds -> 18042
  * Value found -> false
  ****************************************************
  * 
  * 
- * Searching a value not present in the Balanced BST-> 1001
- * Time taken in nano seconds -> 333
+ * Searching value in the Balanced BST-> 1001
+ * Time taken in nano seconds -> 958
  * Value found -> false
  ****************************************************
  * 
@@ -666,56 +301,14 @@ public final class BST {
  * ------------------------------------------------------------------------------
  ****************************************************
  * 
- * Searching first value of input range in Right Skewed BST -> 1
- * Time taken in nano seconds -> 125
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching first value of input range in the Balanced BST-> 1
- * Time taken in nano seconds -> 416
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching middle value of input range in Right Skewed BST -> 5000
- * Time taken in nano seconds -> 49375
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching middle value of input range in the Balanced BST-> 5000
- * Time taken in nano seconds -> 333
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching last value in Right Skewed BST -> 10000
- * Time taken in nano seconds -> 97417
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching last value in balanced BST-> 10000
- * Time taken in nano seconds -> 416
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching a value not present in the Right Skewed BST -> 10001
- * Time taken in nano seconds -> 97125
+ * Searching value in the Right Skewed BST -> 10001
+ * Time taken in nano seconds -> 168833
  * Value found -> false
  ****************************************************
  * 
  * 
- * Searching a value not present in the Balanced BST-> 10001
- * Time taken in nano seconds -> 250
+ * Searching value in the Balanced BST-> 10001
+ * Time taken in nano seconds -> 458
  * Value found -> false
  ****************************************************
  * 
@@ -732,56 +325,14 @@ public final class BST {
  * ------------------------------------------------------------------------------
  ****************************************************
  * 
- * Searching first value of input range in Right Skewed BST -> 1
- * Time taken in nano seconds -> 208
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching first value of input range in the Balanced BST-> 1
- * Time taken in nano seconds -> 583
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching middle value of input range in Right Skewed BST -> 50000
- * Time taken in nano seconds -> 413125
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching middle value of input range in the Balanced BST-> 50000
- * Time taken in nano seconds -> 500
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching last value in Right Skewed BST -> 100000
- * Time taken in nano seconds -> 227250
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching last value in balanced BST-> 100000
- * Time taken in nano seconds -> 625
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching a value not present in the Right Skewed BST -> 100001
- * Time taken in nano seconds -> 230042
+ * Searching value in the Right Skewed BST -> 100001
+ * Time taken in nano seconds -> 1138500
  * Value found -> false
  ****************************************************
  * 
  * 
- * Searching a value not present in the Balanced BST-> 100001
- * Time taken in nano seconds -> 333
+ * Searching value in the Balanced BST-> 100001
+ * Time taken in nano seconds -> 2958
  * Value found -> false
  ****************************************************
  * 
@@ -798,56 +349,14 @@ public final class BST {
  * ------------------------------------------------------------------------------
  ****************************************************
  * 
- * Searching first value of input range in Right Skewed BST -> 1
- * Time taken in nano seconds -> 416
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching first value of input range in the Balanced BST-> 1
- * Time taken in nano seconds -> 1500
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching middle value of input range in Right Skewed BST -> 500000
- * Time taken in nano seconds -> 734625
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching middle value of input range in the Balanced BST-> 500000
- * Time taken in nano seconds -> 5125
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching last value in Right Skewed BST -> 1000000
- * Time taken in nano seconds -> 1880459
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching last value in balanced BST-> 1000000
- * Time taken in nano seconds -> 1875
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching a value not present in the Right Skewed BST -> 1000001
- * Time taken in nano seconds -> 1448583
+ * Searching value in the Right Skewed BST -> 1000001
+ * Time taken in nano seconds -> 3435250
  * Value found -> false
  ****************************************************
  * 
  * 
- * Searching a value not present in the Balanced BST-> 1000001
- * Time taken in nano seconds -> 708
+ * Searching value in the Balanced BST-> 1000001
+ * Time taken in nano seconds -> 1416
  * Value found -> false
  ****************************************************
  * 
@@ -864,56 +373,14 @@ public final class BST {
  * ------------------------------------------------------------------------------
  ****************************************************
  * 
- * Searching first value of input range in Right Skewed BST -> 1
- * Time taken in nano seconds -> 750
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching first value of input range in the Balanced BST-> 1
- * Time taken in nano seconds -> 4167
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching middle value of input range in Right Skewed BST -> 5000000
- * Time taken in nano seconds -> 39010209
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching middle value of input range in the Balanced BST-> 5000000
- * Time taken in nano seconds -> 3541
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching last value in Right Skewed BST -> 10000000
- * Time taken in nano seconds -> 21926750
- * Value found -> true
- ****************************************************
- * 
- * 
- * Searching last value in balanced BST-> 10000000
- * Time taken in nano seconds -> 28500
- * Value found -> true
- ****************************************************
- ****************************************************
- * 
- * 
- * 
- * Searching a value not present in the Right Skewed BST -> 10000001
- * Time taken in nano seconds -> 19606750
+ * Searching value in the Right Skewed BST -> 10000001
+ * Time taken in nano seconds -> 46213833
  * Value found -> false
  ****************************************************
  * 
  * 
- * Searching a value not present in the Balanced BST-> 10000001
- * Time taken in nano seconds -> 3333
+ * Searching value in the Balanced BST-> 10000001
+ * Time taken in nano seconds -> 3791
  * Value found -> false
  ****************************************************
  * 
@@ -923,4 +390,5 @@ public final class BST {
  * Testing completed for skewed and balancedBST of size -> 10000000
  * 
  * ------------------------------------------------------------------------------
- ******************************************************************************/
+ * ozzy@Mac src %
+ ****/
